refactor: split titlebar accessory drag views
