+++ Sources/TitlebarInteractiveControlModifier.swift
@@ -1,63 +1,4 @@
-import AppKit
 import SwiftUI
-
-final class TitlebarAccessoryContainerView: NSView {
-    fileprivate static func shouldResolveWindowDragHit(eventType: NSEvent.EventType?) -> Bool {
-        eventType == nil || eventType == .leftMouseDown
-    }
-
-    override var mouseDownCanMoveWindow: Bool { false }
-
-    override func hitTest(_ point: NSPoint) -> NSView? {
-        guard bounds.contains(point) else { return nil }
-        guard Self.shouldResolveWindowDragHit(eventType: NSApp.currentEvent?.type) else {
-            return super.hitTest(point)
-        }
-        return super.hitTest(point) ?? self
-    }
-
-    override func mouseDown(with event: NSEvent) {
-        if event.clickCount >= 2 {
-            let result = handleTitlebarDoubleClick(
-                window: window,
-                behavior: .standardAction
-            )
-            if result.consumesEvent {
-                return
-            }
-        }
-
-        guard !isWindowDragSuppressed(window: window) else { return }
-
-        if let window {
-            withTemporaryWindowMovableEnabled(window: window) {
-                window.performDrag(with: event)
-            }
-        } else {
-            super.mouseDown(with: event)
-        }
-    }
-}
-
-final class TitlebarAccessoryHostingView<Content: View>: NSHostingView<Content> {
-    override var mouseDownCanMoveWindow: Bool { false }
-
-    override func hitTest(_ point: NSPoint) -> NSView? {
-        guard bounds.contains(point) else { return nil }
-        guard TitlebarAccessoryContainerView.shouldResolveWindowDragHit(eventType: NSApp.currentEvent?.type) else {
-            return super.hitTest(point)
-        }
-        guard let window else { return nil }
-
-        let locationInWindow = convert(point, to: nil)
-        guard isMinimalModeTitlebarControlHit(window: window, locationInWindow: locationInWindow) else {
-            return nil
-        }
-        return super.hitTest(point) ?? self
-    }
-}
-
-typealias NonDraggableHostingView<Content: View> = TitlebarAccessoryHostingView<Content>
 
 /// Protects an interactive control hosted in (or over) the window titlebar from
 /// window-management gestures — window drag, resize drag, and the double-click
